Fixed landing page UI and export pdf issue (#7)

+++ backend/models/stats_model.py
@@ -15,7 +15,7 @@ class RecentCallStat(BaseModel):
     model: str
     inputTokens: int
     outputTokens: int
-    latencyMs: int
+    latencyMs: float
     timestamp: Optional[datetime]
 
 class UserStatsResponse(BaseModel):
